+++ src/main/java/org/opensearch/searchrelevance/common/PluginConstants.java
@@ -34,6 +34,8 @@ public class PluginConstants {
     /** Use %SearchText% to represent wildcard in queryBody and also refer to the text in the search bar */
     public static final String WILDCARD_QUERY_TEXT = "%SearchText%";
 
+    public static final int MAX_TOTAL_FIELDS_LIMIT = 100_000;
+
     /**
      * Indices constants
      */